requirements.txt
langchain==0.0.353
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
altair<5
tiktoken==0.5.2
openai==1.6.1
pymupdf==1.23.26
azure-ai-documentintelligence==1.0.0b2
# langchain-openai==0.1.1